Drop priority from BedrockCommand scheduling
Every command effectively ran at PRIORITY_NORMAL, so the priority
dimension on the scheduling queue was dead weight. The queue now
sorts by scheduled time only, with timeouts still taking precedence.

SScheduledPriorityQueue had a single consumer (BedrockCommandQueue),
so it's been folded directly into BedrockCommandQueue rather than
left as a generic template.

The "priority" field on incoming command requests is now silently
dropped. Jobs.jobPriority (orders rows in the SQLite jobs table) is
unaffected.
